src/requirements.txt
google-generativeai==0.3.2
slack-bolt==1.18.1

